backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-docx
pdfminer.six
openpyxl
pillow
pytesseract
sqlalchemy
psycopg2-binary
nltk
neo4j
python-multipart
faiss-cpu
tqdm
openai
watchdog
